% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addcomma.R, R/initial_functions.R
\name{addcomma}
\alias{addcomma}
\title{Add comma Function}
\usage{
addcomma(x)

addcomma(x)
}
\arguments{
\item{x}{The number you wish to add commas to.}
}
\description{
This function allows you to group every three digits in a number using a comma.

This function allows you to group every three digits in a number using a comma.
}
\examples{
addcomma(23564715565)

addcomma(23564715565)

}
\keyword{comma}
