% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_functions.R
\name{calculate_segment_means}
\alias{calculate_segment_means}
\title{Calculate Segment Means}
\usage{
calculate_segment_means(df)
}
\arguments{
\item{df}{The data frame of segmentation variables that also inclueds a segment column for segment membership}
}
\description{
This function calculates segment means for each segmentation variable
based on the segment memberships uncovered in the cluster analysis.
}
\keyword{averages}
\keyword{calculate}
