% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_console.R
\name{clear_console}
\alias{clear_console}
\title{Clear the Console}
\usage{
clear_console()
}
\description{
This function clears the console. 

In RStudio, Console is located in the lower 

left panel where previously-executed codes

and their outputs are displayed.
}
\examples{
\dontrun{

clear_console()
}

}
