% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyart.R
\name{easyart}
\alias{easyart}
\title{Easy Math Art: Times table on a Circle}
\usage{
easyart(n = 2, mycolor = "deeppink", m = 0)
}
\arguments{
\item{n}{Can be a positive integer, or even a real number - defaults to n = 2.
Choosing n=0 shows 23 interesting patterns.}

\item{mycolor}{Any color name in R. To get color names try colors() in the Console.
defaults to mycolor = "deeppink"}

\item{m}{optional argument that defaults to 0. If a value for it is provided, you will see all the patterns from n=1 upto n=m. For example, setting m=200, shows you all the pattern from 1 to 200")}
}
\description{
This function draws a circle, places 1000 equi-distanced dots on it
and then connects them by a pattern determined by the value of n.
For the interesting math behind it all, watch: https://youtu.be/qhbuKbxJsk8
}
