% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.rlang.R, R/initial_functions.R
\name{fix.rlang}
\alias{fix.rlang}
\title{Fix rlang package}
\usage{
fix.rlang()

fix.rlang()
}
\description{
This function removes and installs rlang to update it to the newest version so that the broom package can be installed

This function removes and installs rlang to update it to the newest version so that the broom package can be installed
}
\examples{
fix.rlang()

fix.rlang()

}
\keyword{rlang}
