% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdashboard.R
\name{getdashboard}
\alias{getdashboard}
\title{Get the Dashboard}
\usage{
getdashboard()
}
\description{
Copy the code after ## Not Run in the Examples below and paste it into a new R Script and save the file as AnalysisHub.R
You can then use this R script to run all of your analyses for this course.
}
\examples{
\dontrun{
getdashboard()
}
}
