% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdata}
\alias{getdata}
\title{Opens a dialog for user to select a csv data file}
\usage{
getdata()
}
\value{

}
\description{
Opens a dialog for user to select a csv data file
}
\examples{
df = getdata()
}
