% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_csv_data.R
\name{load_csv_data}
\alias{load_csv_data}
\title{Open a dialog to select a csv data file}
\usage{
load_csv_data()
}
\description{
This function opens a file selection dialog and
 allows user to select a file. Any file can be 
 selected, but only csv files will be accepted.
 The working directory will also be set to the
 folder that contains the selected file. The working
 directory is the directory where the R session looks for data 
 files or saves the results.
}
\examples{
\dontrun{

df = load_csv_data()
}

}
