% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_functions.R
\name{Plot_MS}
\alias{Plot_MS}
\title{Plot Market Shares in a pie chart}
\usage{
Plot_MS(MS, rule)
}
\arguments{
\item{MS}{Market Share}

\item{rule}{(1 = First Choice, 2 = Preference Share, 3 = Logit)}
}
\description{
Plot Market Shares in a pie chart
}
