% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Conjoint_Analysis.R
\name{Run_Conjoint_Analysis}
\alias{Run_Conjoint_Analysis}
\title{Title}
\usage{
Run_Conjoint_Analysis(
  design,
  products,
  ratings,
  alternatives,
  competitors,
  resizepaper = 1
)
}
\arguments{
\item{design}{the design table csv file}

\item{products}{the list of tested products}

\item{ratings}{ratings of the products in the conjoint survey}

\item{alternatives}{list of proposed product alternatives}

\item{competitors}{list of competitors' products in the market}

\item{resizepaper}{how much larger should the report paper be to fit everything?}
}
\value{
a pdf file
}
\description{
Title
}
\examples{
x=1:10 #just to get rid of the warning
}
