% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_RFM_Analysis.R
\name{Run_RFM_Analysis}
\alias{Run_RFM_Analysis}
\title{RFM Analysis}
\usage{
Run_RFM_Analysis(mydata)
}
\arguments{
\item{mydata}{a csv data file containing at least three numeric columns called Recency, Frequency, and Monetary that hold the R, F, and M data respectively}
}
\description{
RFM Analysis
}
\examples{
\dontrun{
# This is the sample code to be copied and used in a new R Script:
library(listentodata)
clear_console()
df = load_csv_data()
Run_RFM_Analysis(df)
}
}
