% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_SegmentTarget.R
\name{Run_SegmentTarget}
\alias{Run_SegmentTarget}
\title{This function runs a segmentation and targeting analysis and produces results
in a pdf file.}
\usage{
Run_SegmentTarget(df_seg, df_targ, HowManySegments, larger = 1)
}
\arguments{
\item{df_seg}{}

\item{df_targ}{}

\item{HowManySegments}{}

\item{larger}{}
}
\value{
a pdf file
}
\description{
This function runs a segmentation and targeting analysis and produces results
in a pdf file.
}
\examples{
Run_SegmentTarget(segdata, targdata, 3, 1)
}
