% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_folder.R
\name{select_folder}
\alias{select_folder}
\title{Select data folder or results folder and set it as working directory}
\usage{
select_folder()
}
\description{
This function opens a folder selection dialog and
 allows user to select a folder. Then, it sets the
 working directory to the selected directory. The
 working directory is the directory where the R 
 session looks for data files or saves the results.
}
\examples{
\dontrun{

select_folder()
}

}
