% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_CLV_Analysis.R
\name{Run_CLV_Analysis}
\alias{Run_CLV_Analysis}
\title{CLV Analysis}
\usage{
Run_CLV_Analysis(
  profitgroups_data,
  transitionmatrix_data,
  discount_rate,
  new_customers,
  more_customers,
  resizepaper = 1.2
)
}
\arguments{
\item{profitgroups_data}{Profit groups data csv file}

\item{transitionmatrix_data}{Transition matrix data csv file}

\item{discount_rate}{Discount rate}

\item{new_customers}{Number of new customers added on avg. in each period}

\item{more_customers}{Number of extra customers added due to promotions in any of the first 10 periods.}

\item{resizepaper}{How much larger should the pdf paper size be for the results to fit?}
}
\description{
This function conducts a CLV analysis on two csv data files containing data on profit groups and their transition matrix.
}
\examples{
\dontrun{
# This is the sample code to be copied and used in a new R Script:
library(listentodata)
clear_console()
profitgroups_data = load_csv_data()
transitionmatrix_data = load_csv_data()
discount_rate = 0.15
new_customers = 0
more_customers = c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0)
resizepaper = 1.2
Run_CLV_Analysis(profitgroups_data, transitionmatrix_data, discount_rate, new_customers, more_customers, resizepaper)
}
}
