% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_CLV_Analysis.R
\name{Run_CLV_Analysis}
\alias{Run_CLV_Analysis}
\title{CLV Analysis}
\usage{
Run_CLV_Analysis(
  profitgroups_data,
  transitionmatrix_data,
  discount_rate,
  avg_new_customers_each_period,
  extra_customers_period1to10
)
}
\arguments{
\item{profitgroups_data}{}

\item{transitionmatrix_data}{}

\item{discount_rate}{}

\item{avg_new_customers_each_period}{}

\item{extra_customers_period1to10}{}
}
\value{

}
\description{
CLV Analysis
}
