% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Conjoint_Analysis.R
\name{Run_Conjoint_Analysis}
\alias{Run_Conjoint_Analysis}
\title{Title}
\usage{
Run_Conjoint_Analysis(
  design,
  products,
  ratings,
  alternatives,
  competitors,
  resize_page = 1
)
}
\arguments{
\item{design}{}

\item{products}{}

\item{ratings}{}

\item{alternatives}{}

\item{competitors}{}

\item{resize_page}{}
}
\value{

}
\description{
Title
}
\examples{
x=10
}
