% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Conjoint_Study_Design.R
\name{Run_Conjoint_Study_Design}
\alias{Run_Conjoint_Study_Design}
\title{Conjoint Study Design}
\usage{
Run_Conjoint_Study_Design(designparameters)
}
\arguments{
\item{designparameters}{list of attributes and values}
}
\description{
This function takes a set of conjoint design attributes and their values and
returns the set of profiles (hypothetical products) to be used in a conjoint
study survey. The results will be displayed in the Console and also saved in
a csv file.
}
\examples{
\dontrun{
# This is the sample code to be copied and used in a new R Script:
library(listentodata)
clear_console()
designparameters = list(
  NameOfAttribute1 = c("$100","$200","$300","500"),
  NameOfAttribute2 = c("A","B","C","D","E"),
  NameOfAttribute3 = c("Low","Medium","High"),
  NameOfAttribute4 = c("Yes","No"),
  NameOfAttribute5 = c("option1", "option2", "option3", "option4"),
  NameOfAttribute6 = c("option11", "option22", "option33", "option44")
  
)
select_folder()
Run_Conjoint_Study_Design(designparameters) 
}
}
