% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Conjoint_Study_Design.R
\name{Run_Conjoint_Study_Design}
\alias{Run_Conjoint_Study_Design}
\title{Title}
\usage{
Run_Conjoint_Study_Design(designparameters)
}
\arguments{
\item{designparameters}{list of attributes and values}
}
\value{
a csv file and Console output
}
\description{
Title
}
\examples{
x=1:10 #just to get rid of the warning
}
