% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Logistic.R
\name{Run_Logistic}
\alias{Run_Logistic}
\title{Logistic Regression Analysis}
\usage{
Run_Logistic(mydata, myformula, prediction_data = 0, resizepaper = 1)
}
\arguments{
\item{mydata}{csv data file}

\item{myformula}{formula for the glm() model}

\item{prediction_data}{csv data file with the same set of variable names as those in the mydata data set; intended for prediction. The da
The default value set at 0 tells the function that no prediction data set has been provided.}

\item{resizepaper}{How much larger should the pdf paper size be to fit everything?}
}
\description{
This function conducts a Logistic regression analysis and provides the results in a pdf file. 
It can also perform predictions on a new csv data file and provide the results in a csv file.
}
\examples{
\dontrun{
# This is the sample code to be copied and used in a new R Script:
library(listentodata)
clear_console()
mydata = load_csv_data()
head(mydata)
str(mydata)
myformula = "visited ~ ."
resizepaper = 1
Run_Logistic(mydata, myformula)
# If you have data for prediction:
prediction_data = load_csv_data()
Run_Logistic(mydata, myformula, prediction_data)
}
}
