% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Optimal_Pricing.R
\name{Run_Optimal_Pricing}
\alias{Run_Optimal_Pricing}
\title{Optimal Pricing}
\usage{
Run_Optimal_Pricing(
  survey_data,
  prices,
  probabilities,
  marketsize = 10000,
  resizepaper = 1.2
)
}
\arguments{
\item{survey_data}{}

\item{prices}{}

\item{probabilities}{}

\item{resizepaper}{}
}
\value{

}
\description{
Optimal Pricing
}
\examples{
x=10
}
