% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Optimal_Pricing.R
\name{Run_Optimal_Pricing}
\alias{Run_Optimal_Pricing}
\title{Run Optimal Pricing}
\usage{
Run_Optimal_Pricing(
  survey_data,
  prices,
  probabilities,
  marketsize = 10000,
  resizepaper = 1.2
)
}
\arguments{
\item{survey_data}{pricing survey data}

\item{prices}{vector of tested prices}

\item{probabilities}{actual purchase probabilities}

\item{marketsize}{total size of the market}

\item{resizepaper}{How much larger should the pdf paper size be to fit everything?}
}
\value{
a pdf file
}
\description{
Run Optimal Pricing
}
\examples{
x=1:10 #just to get rid of the warning
}
