% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Positioning_Analysis.R
\name{Run_Positioning_Analysis}
\alias{Run_Positioning_Analysis}
\title{Run Positioning Analysis}
\usage{
Run_Positioning_Analysis(percetpions_data, preferences_data)
}
\arguments{
\item{percetpions_data}{Perceptual data}

\item{preferences_data}{Preferences data}
}
\value{
a pdf file
}
\description{
Run Positioning Analysis
}
\examples{
x=1:10 #just to get rid of the warning
}
