% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Positioning_Analysis.R
\name{Run_Positioning_Analysis}
\alias{Run_Positioning_Analysis}
\title{Positioning Analysis}
\usage{
Run_Positioning_Analysis(percetpions_data, preferences_data)
}
\arguments{
\item{preferences_data}{}

\item{perceptions_data}{}
}
\value{

}
\description{
Positioning Analysis
}
\examples{
x=1:10 #just to get rid of the warning
}
