% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Regression.R
\name{Run_Regression}
\alias{Run_Regression}
\title{Title}
\usage{
Run_Regression(mydata, myformula)
}
\arguments{
\item{mydata}{csv data file}

\item{myformula}{formula for the lm model}
}
\value{
a pdf file
}
\description{
Title
}
\examples{
x=1:10 #just to get rid of the warning
}
