% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Regression.R
\name{Run_Regression}
\alias{Run_Regression}
\title{Regression Analysis}
\usage{
Run_Regression(mydata, myformula, newdata)
}
\arguments{
\item{mydata}{csv data file}

\item{myformula}{formula for the lm() model}

\item{newdata}{csv data file without DV intended for prediction}
}
\description{
This function conducts a regression analysis and provides the results in a pdf file.
}
\examples{
\dontrun{
# This is the sample code to be copied and used in a new R Script:
library(listentodata)
clear_console()
library(listentodata)
clear_console()
mydata = load_csv_data()
head(mydata)
str(mydata)
myformula = "Sales ~ ."
or
myformula = "Sales ~ Compprice + Income + Advertising + Population + Price + Age + Education + Urban + US + ShelfLoc"
Run_Regression(mydata, myformula)
}
}
