% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Regression.R
\name{Run_Regression}
\alias{Run_Regression}
\title{Regression Analysis}
\usage{
Run_Regression(mydata, myformula, prediction_data = 0, resizepaper = 1)
}
\arguments{
\item{mydata}{csv data file}

\item{myformula}{formula for the lm() model}

\item{prediction_data}{csv data file with the same set of variable names as those in the mydata data set; intended for prediction. The da
The default value set at 0 tells the function that no prediction data set has been provided by the user.}

\item{resizepaper}{How much larger should the pdf paper size be to fit everything?}
}
\description{
This function conducts a regression analysis and provides the results in a pdf file. 
It can also perform predictions on a new csv data file and provide the results in a csv file.
}
\examples{
\dontrun{
# This is the sample code to be copied and used in a new R Script:
library(listentodata)
clear_console()
mydata = load_csv_data()
head(mydata) #shows top rows in the Console
str(mydata) #shows structure of the data
myformula = "Sales ~ ."
resizepaper = 1
Run_Regression(mydata, myformula, resizepaper)
# If you have data for prediction:
prediction_data = load_csv_data()
Run_Regression(mydata, myformula, prediction_data, resizepaper)
}
}
