% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_SegmentTarget.R
\name{Run_SegmentTarget}
\alias{Run_SegmentTarget}
\title{Segmentation and Targeting Analysis}
\usage{
Run_SegmentTarget(df_seg, df_targ, segments, resizepaper = 1)
}
\arguments{
\item{df_seg}{segmentation data csv file}

\item{df_targ}{targeting data csv file}

\item{segments}{selected number of segments}

\item{resizepaper}{How much larger should the pdf paper size be to fit everything?}
}
\description{
This function conducts a segmentation and targeting analysis on two
segmentation and targeting data sets
}
\examples{
\dontrun{
# This is the sample code to be copied and used in a new R Script:
library(listentodata)
clear_console()
segmentation_data = load_csv_data()
targeting_data = load_csv_data()
segments = 3
resizepaper = 1
Run_SegmentTarget(segmentation_data, targeting_data, segments, resizepaper)
}
}
